+++ src-php/sample-pages-careers.php
@@ -4,7 +4,9 @@
 <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="description" content="Explore current job openings, employee benefits, and career opportunities with the city government." />
+    <title>City Careers - RFPMart + USWDS</title>
+    <meta name="description" content="Explore career opportunities with the city, including job openings, benefits, and how to apply." />
+
 
     <?php require_once '1_head.php'; ?>
 </head>
@@ -19,7 +21,7 @@
                     <nav class="usa-breadcrumb bg-transparent" aria-label="Breadcrumbs">
                         <ul class="usa-breadcrumb__list">
                             <li class="usa-breadcrumb__list-item">
-                                <a href="index.html" class="usa-breadcrumb__link">
+                                <a href="./" class="usa-breadcrumb__link">
                                     <span class="text-base-lightest hover:text-base-light">Home</span>
                                 </a>
                             </li>