+++ src-php/sample-pages-contact-form.php
@@ -4,8 +4,8 @@
 <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="description" content="Reach city departments, submit questions, or request services through our main contact page." />
-
+    <title>Send a Message - RFPMart + USWDS</title>
+    <meta name="description" content="Submit questions or requests directly to the city using the online contact form." />
 
     <?php require_once '1_head.php'; ?>
 </head>
@@ -24,7 +24,7 @@
                             <nav class="usa-breadcrumb bg-transparent" aria-label="Breadcrumbs">
                                 <ul class="usa-breadcrumb__list">
                                     <li class="usa-breadcrumb__list-item">
-                                        <a href="index.html" class="usa-breadcrumb__link">
+                                        <a href="./" class="usa-breadcrumb__link">
                                             <span class="">Home</span>
                                         </a>
                                     </li>