+++ src-php/user-log-in.php
@@ -4,6 +4,8 @@
 <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Log In - RFPMart + USWDS</title>
+    <meta name="description" content="Securely log in to your account to access personalized city services and tools." />
 
     <?php require_once '1_head.php'; ?>
 </head>